% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.ROC.R
\name{auc.ROC}
\alias{auc.ROC}
\title{acu.ROC}
\usage{
auc.ROC(file)
}
\arguments{
\item{file}{The connection to the data in the Univariate file.}
}
\value{
A matrix to outline the AUC and the confidence region.
}
\description{
This function provide the receiver operating characteristic curve and Area under the Curve of ROC.
}

