% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.plot.R
\name{box.plot}
\alias{box.plot}
\title{Box Plot}
\usage{
box.plot(file)
}
\arguments{
\item{file}{The connection for intensity of metabolites in each group.}
}
\value{
The box plot
}
\description{
Box Plot of each metabolite between groups.
}

