% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bStatX.R
\name{bStatX}
\alias{bStatX}
\title{Description statistic}
\usage{
bStatX(file)
}
\arguments{
\item{file}{The file with the expression information.}
}
\value{
A matrix for data description
}
\description{
bStatX provide the basic data description for each features. Data descriptions includes mean value, median value, sum, quartile, standard derivatives, etc.
}

