% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTarget_rforest.R
\name{mdsPlot}
\alias{mdsPlot}
\title{MDSplot in statTarget}
\usage{
mdsPlot(rForest,pimpModel,Labels = TRUE,slink = FALSE, 
slinkDat, ...)
}
\arguments{
\item{rForest}{An object of class randomForest that contains the proximity 
component from statTarget_rForest function.}

\item{pimpModel}{An object of permutation-based variable Gini importance 
measures (PIMP-algorithm) from statTarget_rForest function.}

\item{Labels}{Labels is TRUE for visible the sample name in the figure else 
with the index for class.}

\item{slink}{Logical indicating if slinkDat is active for extenal classID.}

\item{slinkDat}{A data frame for the extenal classID.}

\item{...}{A generic MDSplot function in randomForest package}
}
\value{
The output of cmdscale on 1 - rf$proximity is returned invisibly.
}
\description{
Multi-dimensional scaling plot of proximity matrix from 
randomForest.
}
\examples{
datpath <- system.file('extdata',package = 'statTarget')
statFile <- paste(datpath,'data_example.csv', sep='/')
getFile <- read.csv(statFile,header=TRUE)
rFtest <- rForest(getFile,ntree = 10,times = 5)
mdsPlot(rFtest$randomForest,rFtest$pimpTest)
}
\seealso{
MDSplot
}
\author{
Hemi Luan, hemi.luan@gmail.com
}
