% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.stat.R
\name{plsda.stat}
\alias{plsda.stat}
\title{oplsda}
\usage{
plsda.stat(scaling, silt)
}
\arguments{
\item{scaling}{The Scaling method before statistic analysis (PCA or PLS-DA). 'Pareto' can be used for specifying the Pareto scaling. 'Auto' can be used for specifying the Auto scaling (or unit variance scaling). 'Vast' can be used for specifying the vast scaling. 'Range' can be used for specifying the Range scaling.}

\item{silt}{The permutation time in PLS-DA model for single-response models!}
}
\value{
The matrix of PLS-DA model parameters.
}
\description{
This function provide PLS-DA model according to the class subgroups.
}

