% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_RF.R
\name{predict_RF}
\alias{predict_RF}
\title{preidict function for random forest objects in statTarget}
\usage{
predict_RF(object, newdata, type='response',...)
}
\arguments{
\item{object}{An object created by the function statTarget_rForest.}

\item{newdata}{A data frame or matrix containing new data. (Note: If not 
given, the out-of-bag prediction in object is returned. see randomForest 
package.}

\item{type}{One of response, prob. or votes, indicating the type of output: 
predicted values, matrix of class probabilities, or matrix of vote counts. 
class is allowed, but automatically converted to 'response', for backward 
compatibility.}

\item{...}{A generic predict function from randomForest package.}
}
\value{
A class of predicted values is returned. 
Object type is classification, for detail see randomForest package.
}
\description{
Prediction of test data using random forest in statTarget.
}
\seealso{
randomForest
}
\author{
Hemi Luan, hemi.luan@gmail.com
}
