% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{pvalues}
\alias{pvalues}
\title{P values}
\usage{
pvalues(file, fdr)
}
\arguments{
\item{file}{The connection to the data in the Univariate file.}

\item{fdr}{The false discovery rate for conceptualizing the rate of type I errors in null hypothesis testing when conducting multiple comparisons.}
}
\value{
A matrix with p values
}
\description{
This function displyed the  adjusted p values with BH method for each pairwise comparison of sample groups, by intergrating the p values from Student T test for normally distributed variables and from Wilcoxon Mann Whitney test for non-normally distributed variables.
}

