% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro.R
\name{shapiro}
\alias{shapiro}
\title{Shapiro test}
\usage{
shapiro(file)
}
\arguments{
\item{file}{The connection to the data file.}
}
\value{
A result with Shapiro Wilk test
}
\description{
This function provide the Shapiro Wilk test is a test of normality of each pairwise comparison of sample groups.
}

