% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statAnalysis.R
\name{statAnalysis}
\alias{statAnalysis}
\title{statAnalysis provide the statistical analysis for metabolomics data or others.}
\usage{
statAnalysis(file, Frule = 0.8, imputeM = "KNN", glog = TRUE,
  test.multi = TRUE, nvarRF = 10, scaling = "Pareto", silt = 500,
  pcax = 1, pcay = 2)
}
\arguments{
\item{file}{a file with  the expression information.}

\item{Frule}{The cut-off value for missing value filter function.}

\item{imputeM}{The parameter for imputation method.(i.e., nearest neighbor averaging, "KNN"; minimum values for imputed variables, "min", median values for imputed variables (Group dependent) "median").}

\item{glog}{is a data index for log transformation, with the default value TRUE.}

\item{test.multi}{must be an index for statistic analysis, with the default value TRUE.}

\item{nvarRF}{shows the number of variables in Gini plot of Randomforest model (=< 100).}

\item{scaling}{is index of scaling method index for statistic analysis (PCA or PLS-DA). 'pareto', 'Pareto', 'p' or 'P' can be used for specifying the Pareto scaling. 'auto', 'Auto', 'auto', 'a' or 'A' can be used for specifying the Auto scaling (or unit variance scaling). 'vast', 'Vast', 'v' or 'V' can be used for specifying the vast scaling. 'range', 'Range', 'r' or 'R' can be used for specifying the Range scaling.}

\item{silt}{shows the number of permutation times}

\item{pcax}{shows principal components in PCA model for the x-axis.}

\item{pcay}{shows principal components in PCA model for the y-axis.}
}
\description{
statAnalysis provide the statistical analysis for metabolomics data or others.
}

