% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTarget-package.R
\docType{package}
\name{statTarget-package}
\alias{statTarget}
\alias{statTarget-package}
\title{Statistical Analysis of Metabolite Profile}
\usage{
statTarget()
}
\description{
A graphical user interface, easy to use tool provide quality control based signal correction, integration of metabolomic data from multiple batches, and the comprehensive statistic analysis for non-targeted and targeted approaches.
}
\details{
Package: statTarget

Type: package

Version: 1.2.0

Date: 2016-07-014

License: GPL (>= 2)
}
\author{
Hemi Luan

Maintainer: Hemi Luan hemi.luan@gmail.com
}

