% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transX.R
\name{transCode}
\alias{transCode}
\title{transCode for statTarget inputs}
\usage{
transCode(data, type)
}
\arguments{
\item{data}{A transCode objects. The output file from Mass Spectrometry Data 
softwares.}

\item{type}{The output file formats from Mass Spectrometry Data software, 
including "XCMS" or "xcms",...; The softwares include XCMS (.tsv file 
generated from diffreport function), ...}
}
\value{
A list of inputs
}
\description{
transCode is to generate statTarget inputs from Mass Spectrometry 
Data softwares, like XCMS.
}
\examples{
datpath <- system.file("extdata",package = "statTarget")
data <- paste(datpath,"xcmsOutput.tsv", sep="/")
transCode(data,"xcms")
}
\keyword{XCMS}
\keyword{inputs}

