% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{Volcano Plot}
\usage{
volcano(file, upper.lim, lower.lim, sig.lim)
}
\arguments{
\item{file}{The connection for P-values and Fold changes of metabolites.}

\item{upper.lim}{The up-regulated metabolites using Fold Changes cut off values in the Volcano plot.}

\item{lower.lim}{The down-regulated metabolites using Fold Changes cut off values in the Volcano plot.}

\item{sig.lim}{The significance level for metabolites in the Volcano plot.}
}
\value{
The volcano plot
}
\description{
volcano provide the differential markers for metabolomics analysis.
}

