% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch.R
\name{welch}
\alias{welch}
\title{Student's T-tests}
\usage{
welch(file)
}
\arguments{
\item{file}{The connection to the data file.}
}
\value{
A matrix with p value
}
\description{
welch This function provide two.sided Student's T-tests for each pairwise comparison of sample groups.
}

