% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmw.R
\name{wmw}
\alias{wmw}
\title{Wilcoxon-Mann Whitney tests}
\usage{
wmw(file)
}
\arguments{
\item{file}{The connection to the data file.}
}
\value{
A matrix with p values
}
\description{
This function provide two.sided Wilcoxon Mann Whitney tests for each pairwise comparison of sample groups.
}

