% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bStatX.R
\name{bStatX}
\alias{bStatX}
\title{bStatX provide the basic data description for each features}
\usage{
bStatX(file)
}
\arguments{
\item{file}{is the data input}
}
\description{
bStatX provide the basic data description for each features
}

