% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missvalue.R
\name{missvalue}
\alias{missvalue}
\title{Miss value will be replaced with median value.}
\usage{
missvalue(mv, degree)
}
\arguments{
\item{mv}{is th data file}

\item{degree}{is the group factor}
}
\description{
Miss value will be replaced with median value.
}

