% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd.ratio.R
\name{odd.ratio}
\alias{odd.ratio}
\title{this function provide the odd.ratio value for metabolomics data or others.}
\usage{
odd.ratio(file)
}
\arguments{
\item{file}{is the data imput}
}
\description{
this function provide the odd.ratio value for metabolomics data or others.
}
\details{
An odds ratio (OR) is a measure of association between an exposure and an outcome. OR=1 Exposure does not affect odds of outcome; OR>1 Exposure associated with higher odds of outcome; OR<1 Exposure associated with lower odds of outcome.The 95% confidence interval (CI) is used to estimate the precision of the OR. A large CI indicates a low level of precision of the OR, whereas a small CI indicates a higher precision of the OR. It is important to note however, that unlike the p value, the 95% CI does not report a measure???s statistical significance.
}
\references{
J Can Acad Child Adolesc Psychiatry. 2010 Aug; 19(3): 227???229. Magdalena Szumilas
}

