% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomF.R
\name{RandomF}
\alias{RandomF}
\title{RandomF provide the random forest function.}
\usage{
RandomF(file, nvarRF)
}
\arguments{
\item{file}{is the data}

\item{nvarRF}{shows the number of variables in Gini plot of Randomforest model (=< 100).}
}
\description{
Multi result was outputed, such as MDS plot, vriable Gini plot and list results
}

