% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTarget-package.R
\docType{package}
\name{statTarget-package}
\alias{statTarget-package}
\alias{statTarget}
\title{Statistical Analysis of Molecular Profiles}
\usage{
statTarget()
}
\value{
A description of statTarget. See the details at https://stattarget.github.io
}
\description{
An streamlined tool provides graphical user interface 
for quality control based signal correction, integration of 
MS-based data from multiple batches, and the 
comprehensive statistical analysis for omics studies.
}
\details{
Package: statTarget

Type: package

License: LGPL (>= 3)
}
\author{
Hemi Luan

Maintainer: Hemi Luan hemi.luan@gmail.com
}
