% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTarget-package.R
\docType{package}
\name{statTarget-package}
\alias{statTarget}
\alias{statTarget-package}
\title{Statistical Analysis of Metabolite Profile}
\usage{
statTarget()
}
\value{
A description of statTarget
}
\description{
An easy to use tool provide graphical user interface 
for quality control based signal correction, integration of 
metabolomic data from multiple batches, and the 
comprehensive statistic analysis for non-targeted and targeted approaches.
}
\details{
Package: statTarget

Type: package

Version: 1.2.4

Date: 2016-07-014

License: GPL (>= 2)
}
\author{
Hemi Luan

Maintainer: Hemi Luan hemi.luan@gmail.com
}

