% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTargetGUI.R
\name{StatTargetGUI}
\alias{StatTargetGUI}
\alias{statTargetGUI}
\title{StatTargetGUI for GUI}
\usage{
statTargetGUI()
}
\value{
The output of GUI
}
\description{
the statTarget GUI session. The Shift Correction and 
Statistical Analysis session being used by statTarget. Will restart 
statTarget if it died for some reason. Features of the package statTarget 
includes shift correction, typical quality control based robust LOESS 
signal correction (such as QC.RLSC); Data preprocessing, data descriptions,
PCA, PLSDA, OPLSDA, VIP, ROC, random forest, odd ratio, Student T-test, 
Shapiro-Wilk normality test and Mann-Whitney tests; Data preprocessing 
includes 80-precent rule, log transformation, normalization. 
Data descriptions includes mean value, median value, sum, quartile, 
standard derivatives, etc.
}
\examples{
if (interactive()) {statTargetGUI()}
}
\author{
Hemi Luan hemi.luan@gmail.com
}
\references{
Dunn WB., et al. Nat Protoc. 2011, 6, pp1060.
Luan H., et al. GigaScience 2015, 4, pp16.
Luan H., et al. J. Proteome Res., 2015, 14, pp467.
}
\keyword{A}
\keyword{GUI}
\keyword{QCRLSC}
\keyword{analysis}
\keyword{correction}
\keyword{for}
\keyword{metabolomics}
\keyword{of}
\keyword{shift}
\keyword{statTarget}
\keyword{statistical}

