% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transX.R
\name{transX}
\alias{transX}
\title{transX for statTarget inputs}
\usage{
transX(data, type)
}
\arguments{
\item{data}{A transX objects. The output file from Mass Spectrometry Data 
softwares.}

\item{type}{The output file formats from Mass Spectrometry Data software, 
including "XCMS" or "xcms",...; The softwares include XCMS (.tsv file 
generated from diffreport function), ...}
}
\value{
A objects of transX
}
\description{
transX is to generate statTarget inputs from Mass Spectrometry 
Data softwares, like XCMS.
}
\examples{
datpath <- system.file("extdata",package = "statTarget")
data <- paste(datpath,"xcmsOutput.tsv", sep="/")
transX(data,"xcms")
transCode(data,"xcms")
}
\keyword{XCMS}
\keyword{inputs}

